feat: FriendProvider + 熔断器模式 + 后端warning修复
- Flutter: 新增FriendProvider(好友/临时好友/请求+WebSocket实时更新)
- Flutter: CircuitBreaker熔断器(halfOpen探测+自动恢复)
- Flutter: ResilientApiClient弹性客户端(独立熔断器+缓存降级)
- Flutter: main.dart注册StoneProvider和FriendProvider
- Backend: 修复EventHandlers.h未使用参数、SemanticCache.h类型转换
- 三端零error零warning

+++ backend/include/infrastructure/ai/SemanticCache.h
@@ -31,7 +31,7 @@ struct SemanticCacheStats {
     size_t misses = 0;
     float hitRate() const {
         size_t total = exactHits + semanticHits + misses;
-        return total > 0 ? float(exactHits + semanticHits) / total : 0.0f;
+        return total > 0 ? static_cast<float>(exactHits + semanticHits) / static_cast<float>(total) : 0.0f;
     }
     size_t estimatedSavings() const { return (exactHits + semanticHits) * 2; }
 };
